스키마, 라우터, 모델 리팩토링
