+++ django/app/database/__init__.py
@@ -1,17 +1,15 @@
 # 데이터베이스 연결 모듈
-from .mongodb import mongodb_manager, company_model, company_review_model
+from .mongodb import mongodb_manager
 from .redis_client import redis_client
-from .postgres_models import Inquiry
+from .postgres import tortoise_manager
 
 __all__ = [
-    # MongoDB
-    'mongodb_manager',
-    'company_model', 
-    'company_review_model',
-    
-    # Redis
-    'redis_client',
-    
-    # PostgreSQL (Tortoise ORM)
-    'Inquiry',
+  # MongoDB 연결
+  'mongodb_manager',
+  
+  # Redis 연결
+  'redis_client',
+  
+  # PostgreSQL 연결
+  'tortoise_manager',
 ]
